changed polygon file from >3m acres to 50k acres size
